md_backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
# django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
psycopg2==2.9.3
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
